chore: move seeds config to dedicated file
