+++ api/src/shared/config.js
@@ -82,39 +82,6 @@ function isEnabledByContainerModulo(envVarValue) {
   return containerIndex % modulo === 0;
 }
 
-function getSeedsConfig() {
-  const context = buildSeedsContext(process.env.SEEDS_CONTEXT);
-
-  const frameworks = process.env.SEEDS_LEARNING_CONTENT_FRAMEWORKS?.split(',') ?? ['Pix', 'Droit', 'Edu', 'Modulix'];
-  if (context.junior && !frameworks.includes('Pix 1D')) {
-    frameworks.push('Pix 1D');
-  }
-
-  return {
-    context,
-    learningContent: {
-      frameworks,
-      locales: process.env.SEEDS_LEARNING_CONTENT_LOCALES?.split(',') ?? ['fr-fr', 'en', 'nl', 'nl-BE'],
-    },
-  };
-}
-
-const SEEDS_CONTEXTS = [
-  'prescription',
-  'devcomp',
-  'junior',
-  'acces',
-  'contenu',
-  'certification',
-  'evaluation',
-  'acquisition',
-];
-
-function buildSeedsContext(value) {
-  const values = value && value.length ? value.toLowerCase().split('|') : SEEDS_CONTEXTS;
-  return Object.fromEntries(Array.from(SEEDS_CONTEXTS, (v) => [v, values.includes(v)]));
-}
-
 export const schema = Joi.object({
   MADDO: Joi.boolean().optional().default(false),
   ACCESS_TOKEN_LIFESPAN: Joi.string().optional(),
@@ -565,7 +532,6 @@ export const config = {
     },
     accessTokenLifespanMs: ms(process.env.SAML_ACCESS_TOKEN_LIFESPAN || '7d'),
   },
-  seeds: getSeedsConfig(),
   passwordResetDemand: {
     secret: process.env.AUTH_SECRET,
     lifespan: process.env.PASSWORD_RESET_DEMAND_LIFESPAN || '1h',
